backend/requirements.txt
django>=4.2,<5.0
psycopg2-binary
gunicorn
python-dotenv
django-environ
djangorestframework
django-cors-headers
# GIS support is built into Django, but we need the DB drivers
psycopg2-binary
Pillow>=10.0.0
